feat(sync): Rework geozone sync
The `ogr2ogr` command has been replaced by the updated sync lambda.
Extra fields of geozone entity has been removed to the benefit of
simplicity and at the expense of debugging capability.
We assume that a geozone is never updated, instead we receive a new
one with its version incremented.